Revert "Merge branch 'CELEBORN-360' into CELEBORN-363"
This reverts commit 8626406605ee33a06ab53e349b24d771f4d4aa15, reversing
changes made to 39c6744760d23c1f4c3bbaee7e0cfa2fc6ca1df3.
