--- common/src/main/java/org/apache/celeborn/common/network/server/memory/WrappedDataBuffer.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.celeborn.common.network.server.memory;
-
-import io.netty.buffer.ByteBuf;
-
-public class WrappedDataBuffer {
-
-  public final ByteBuf byteBuf;
-
-  public final Recycler bufferRecycler;
-
-  public WrappedDataBuffer(ByteBuf byteBuf, Recycler bufferRecycler) {
-    this.byteBuf = byteBuf;
-    this.bufferRecycler = bufferRecycler;
-  }
-
-  public void recycle() {
-    bufferRecycler.recycle(byteBuf);
-  }
-
-  public void release() {
-    bufferRecycler.release(byteBuf);
-  }
-}
