--- conf/celeborn-ratis-shell-env.sh.template
@@ -1,1 +1,0 @@
-
